third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/update-text-of-style-in-shadow-dom-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
Tests update the text of styles updates styles in shadow dom, crbug.com/247280

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
